query/planner/table -> sql/planner/plans/ddl/table
