+++ src/query/planner/src/plans/mod.rs
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod alter_table_cluster_key;
 mod alter_udf;
 mod alter_user;
 mod alter_view;
@@ -23,33 +22,23 @@ mod create_udf;
 mod create_user;
 mod create_view;
 mod delete;
-mod describe_table;
 mod drop_role;
 mod drop_stage;
-mod drop_table;
-mod drop_table_cluster_key;
 mod drop_udf;
 mod drop_user;
 mod drop_view;
-mod exists_table;
 mod grant_privilege;
 mod grant_role;
 mod list;
-mod optimize_table;
 mod projection;
 mod remove_stage;
-mod rename_table;
 mod revoke_privilege;
 mod revoke_role;
 mod set_role;
 mod setting;
-mod show_create_table;
 mod show_grants;
 mod show_roles;
-mod truncate_table;
-mod undrop_table;
 
-pub use alter_table_cluster_key::AlterTableClusterKeyPlan;
 pub use alter_udf::AlterUDFPlan;
 pub use alter_user::AlterUserPlan;
 pub use alter_view::AlterViewPlan;
@@ -60,30 +49,19 @@ pub use create_udf::CreateUDFPlan;
 pub use create_user::CreateUserPlan;
 pub use create_view::CreateViewPlan;
 pub use delete::*;
-pub use describe_table::DescribeTablePlan;
 pub use drop_role::DropRolePlan;
 pub use drop_stage::DropStagePlan;
-pub use drop_table::DropTablePlan;
-pub use drop_table_cluster_key::DropTableClusterKeyPlan;
 pub use drop_udf::DropUDFPlan;
 pub use drop_user::DropUserPlan;
 pub use drop_view::DropViewPlan;
-pub use exists_table::ExistsTablePlan;
 pub use grant_privilege::GrantPrivilegePlan;
 pub use grant_role::GrantRolePlan;
 pub use list::ListPlan;
-pub use optimize_table::OptimizeTableAction;
-pub use optimize_table::OptimizeTablePlan;
 pub use projection::*;
 pub use remove_stage::RemoveStagePlan;
-pub use rename_table::RenameTableEntity;
-pub use rename_table::RenameTablePlan;
 pub use revoke_privilege::RevokePrivilegePlan;
 pub use revoke_role::RevokeRolePlan;
 pub use set_role::SetRolePlan;
 pub use setting::*;
-pub use show_create_table::ShowCreateTablePlan;
 pub use show_grants::ShowGrantsPlan;
 pub use show_roles::ShowRolesPlan;
-pub use truncate_table::TruncateTablePlan;
-pub use undrop_table::UndropTablePlan;
