--- src/query/planner/src/plans/truncate_table.rs
@@ -1,32 +1,0 @@
-// Copyright 2022 Datafuse Labs.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-use std::sync::Arc;
-
-use common_datavalues::DataSchema;
-use common_datavalues::DataSchemaRef;
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct TruncateTablePlan {
-    pub catalog: String,
-    pub database: String,
-    /// The table name
-    pub table: String,
-    pub purge: bool,
-}
-
-impl TruncateTablePlan {
-    pub fn schema(&self) -> DataSchemaRef {
-        Arc::new(DataSchema::empty())
-    }
-}
