+++ src/query/service/src/interpreters/interpreter_cluster_key_alter.rs
@@ -15,7 +15,7 @@
 use std::sync::Arc;
 
 use common_exception::Result;
-use common_planner::plans::AlterTableClusterKeyPlan;
+use common_sql::plans::AlterTableClusterKeyPlan;
 
 use super::Interpreter;
 use crate::pipelines::PipelineBuildResult;
