+++ src/query/service/src/interpreters/interpreter_table_rename.rs
@@ -17,7 +17,7 @@ use std::sync::Arc;
 use common_exception::Result;
 use common_meta_app::schema::RenameTableReq;
 use common_meta_app::schema::TableNameIdent;
-use common_planner::plans::RenameTablePlan;
+use common_sql::plans::RenameTablePlan;
 
 use crate::interpreters::Interpreter;
 use crate::pipelines::PipelineBuildResult;
