+++ src/query/service/src/interpreters/interpreter_table_truncate.rs
@@ -15,7 +15,7 @@
 use std::sync::Arc;
 
 use common_exception::Result;
-use common_planner::plans::TruncateTablePlan;
+use common_sql::plans::TruncateTablePlan;
 
 use crate::interpreters::Interpreter;
 use crate::pipelines::PipelineBuildResult;
