+++ src/query/service/tests/it/storages/fuse/pruning.rs
@@ -25,6 +25,7 @@ use common_sql::executor::col;
 use common_sql::executor::lit;
 use common_sql::executor::sub;
 use common_sql::executor::ExpressionOp;
+use common_sql::plans::CreateTablePlanV2;
 use common_storages_fuse::FuseTable;
 use common_storages_table_meta::meta::BlockMeta;
 use common_storages_table_meta::meta::TableSnapshot;
@@ -34,7 +35,6 @@ use databend_query::interpreters::CreateTableInterpreterV2;
 use databend_query::interpreters::Interpreter;
 use databend_query::sessions::QueryContext;
 use databend_query::sessions::TableContext;
-use databend_query::sql::plans::create_table_v2::CreateTablePlanV2;
 use databend_query::storages::fuse::io::MetaReaders;
 use databend_query::storages::fuse::pruning::BlockPruner;
 use databend_query::storages::fuse::FUSE_OPT_KEY_BLOCK_PER_SEGMENT;
