+++ src/query/service/tests/it/storages/fuse/table.rs
@@ -18,11 +18,12 @@ use common_ast::ast::Engine;
 use common_base::base::tokio;
 use common_exception::Result;
 use common_meta_app::schema::TableInfo;
-use common_planner::plans::AlterTableClusterKeyPlan;
-use common_planner::plans::DropTableClusterKeyPlan;
 use common_planner::ReadDataSourcePlan;
 use common_planner::SourceInfo;
 use common_sql::executor::table_read_plan::ToReadDataSourcePlan;
+use common_sql::plans::AlterTableClusterKeyPlan;
+use common_sql::plans::CreateTablePlanV2;
+use common_sql::plans::DropTableClusterKeyPlan;
 use common_storages_table_meta::table::OPT_KEY_DATABASE_ID;
 use common_storages_table_meta::table::OPT_KEY_SNAPSHOT_LOCATION;
 use databend_query::interpreters::AlterTableClusterKeyInterpreter;
@@ -31,7 +32,6 @@ use databend_query::interpreters::DropTableClusterKeyInterpreter;
 use databend_query::interpreters::Interpreter;
 use databend_query::interpreters::InterpreterFactory;
 use databend_query::sessions::TableContext;
-use databend_query::sql::plans::create_table_v2::CreateTablePlanV2;
 use databend_query::sql::Planner;
 use databend_query::storages::fuse::io::MetaReaders;
 use databend_query::storages::fuse::FuseTable;
