+++ src/query/service/tests/it/storages/fuse/table_test_fixture.rs
@@ -27,6 +27,7 @@ use common_meta_app::schema::DatabaseMeta;
 use common_planner::extras::Extras;
 use common_sql::executor::table_read_plan::ToReadDataSourcePlan;
 use common_sql::plans::CreateDatabasePlan;
+use common_sql::plans::CreateTablePlanV2;
 use common_storage::StorageFsConfig;
 use common_storage::StorageParams;
 use common_storages_fuse::FUSE_TBL_XOR_BLOOM_INDEX_PREFIX;
@@ -41,7 +42,6 @@ use databend_query::pipelines::processors::BlocksSource;
 use databend_query::pipelines::PipelineBuildResult;
 use databend_query::sessions::QueryContext;
 use databend_query::sessions::TableContext;
-use databend_query::sql::plans::create_table_v2::CreateTablePlanV2;
 use databend_query::sql::Planner;
 use databend_query::storages::fuse::table_functions::ClusteringInformationTable;
 use databend_query::storages::fuse::table_functions::FuseSnapshotTable;
