+++ src/query/service/tests/it/storages/null.rs
@@ -18,8 +18,8 @@ use common_exception::Result;
 use common_meta_app::schema::TableInfo;
 use common_meta_app::schema::TableMeta;
 use common_sql::executor::table_read_plan::ToReadDataSourcePlan;
+use common_sql::plans::TableOptions;
 use common_storages_null::NullTable;
-use databend_query::sql::plans::create_table_v2::TableOptions;
 use databend_query::stream::ReadDataBlockStream;
 use futures::TryStreamExt;
