+++ src/query/sql/src/planner/binder/ddl/table.rs
@@ -17,8 +17,28 @@ use std::collections::BTreeMap;
 use std::collections::HashSet;
 use std::sync::Arc;
 
+use common_ast::ast::AlterTableAction;
+use common_ast::ast::AlterTableStmt;
+use common_ast::ast::CompactTarget;
+use common_ast::ast::CreateTableSource;
+use common_ast::ast::CreateTableStmt;
+use common_ast::ast::DescribeTableStmt;
+use common_ast::ast::DropTableStmt;
+use common_ast::ast::Engine;
+use common_ast::ast::ExistsTableStmt;
+use common_ast::ast::Expr;
+use common_ast::ast::Literal;
 use common_ast::ast::OptimizeTableAction as AstOptimizeTableAction;
-use common_ast::ast::*;
+use common_ast::ast::OptimizeTableStmt;
+use common_ast::ast::RenameTableStmt;
+use common_ast::ast::ShowCreateTableStmt;
+use common_ast::ast::ShowLimit;
+use common_ast::ast::ShowTablesStatusStmt;
+use common_ast::ast::ShowTablesStmt;
+use common_ast::ast::Statement;
+use common_ast::ast::TableReference;
+use common_ast::ast::TruncateTableStmt;
+use common_ast::ast::UndropTableStmt;
 use common_ast::parser::parse_sql;
 use common_ast::parser::tokenize_sql;
 use common_ast::walk_expr_mut;
@@ -33,18 +53,6 @@ use common_datavalues::TypeFactory;
 use common_datavalues::Vu8;
 use common_exception::ErrorCode;
 use common_exception::Result;
-use common_planner::plans::AlterTableClusterKeyPlan;
-use common_planner::plans::DescribeTablePlan;
-use common_planner::plans::DropTableClusterKeyPlan;
-use common_planner::plans::DropTablePlan;
-use common_planner::plans::ExistsTablePlan;
-use common_planner::plans::OptimizeTableAction;
-use common_planner::plans::OptimizeTablePlan;
-use common_planner::plans::RenameTableEntity;
-use common_planner::plans::RenameTablePlan;
-use common_planner::plans::ShowCreateTablePlan;
-use common_planner::plans::TruncateTablePlan;
-use common_planner::plans::UndropTablePlan;
 use common_storage::parse_uri_location;
 use common_storage::DataOperator;
 use common_storage::UriLocation;
@@ -60,12 +68,24 @@ use crate::optimizer::OptimizerConfig;
 use crate::optimizer::OptimizerContext;
 use crate::planner::semantic::normalize_identifier;
 use crate::planner::semantic::IdentifierNormalizer;
-use crate::plans::create_table_v2::CreateTablePlanV2;
+use crate::plans::AlterTableClusterKeyPlan;
 use crate::plans::CastExpr;
+use crate::plans::CreateTablePlanV2;
+use crate::plans::DescribeTablePlan;
+use crate::plans::DropTableClusterKeyPlan;
+use crate::plans::DropTablePlan;
+use crate::plans::ExistsTablePlan;
+use crate::plans::OptimizeTableAction;
+use crate::plans::OptimizeTablePlan;
 use crate::plans::Plan;
 use crate::plans::ReclusterTablePlan;
+use crate::plans::RenameTableEntity;
+use crate::plans::RenameTablePlan;
 use crate::plans::RewriteKind;
 use crate::plans::Scalar;
+use crate::plans::ShowCreateTablePlan;
+use crate::plans::TruncateTablePlan;
+use crate::plans::UndropTablePlan;
 use crate::BindContext;
 use crate::ColumnBinding;
 use crate::ScalarExpr;
