+++ src/query/sql/src/planner/plans/ddl/mod.rs
@@ -12,5 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 mod database;
+mod table;
 
 pub use database::*;
+pub use table::*;
