+++ src/query/sql/src/planner/plans/plan.rs
@@ -21,7 +21,6 @@ use common_datavalues::DataSchema;
 use common_datavalues::DataSchemaRef;
 use common_datavalues::DataSchemaRefExt;
 use common_datavalues::StringType;
-use common_planner::plans::AlterTableClusterKeyPlan;
 use common_planner::plans::AlterUDFPlan;
 use common_planner::plans::AlterUserPlan;
 use common_planner::plans::AlterViewPlan;
@@ -32,34 +31,24 @@ use common_planner::plans::CreateUDFPlan;
 use common_planner::plans::CreateUserPlan;
 use common_planner::plans::CreateViewPlan;
 use common_planner::plans::DeletePlan;
-use common_planner::plans::DescribeTablePlan;
 use common_planner::plans::DropRolePlan;
 use common_planner::plans::DropStagePlan;
-use common_planner::plans::DropTableClusterKeyPlan;
-use common_planner::plans::DropTablePlan;
 use common_planner::plans::DropUDFPlan;
 use common_planner::plans::DropUserPlan;
 use common_planner::plans::DropViewPlan;
-use common_planner::plans::ExistsTablePlan;
 use common_planner::plans::GrantPrivilegePlan;
 use common_planner::plans::GrantRolePlan;
 use common_planner::plans::ListPlan;
-use common_planner::plans::OptimizeTablePlan;
 use common_planner::plans::RemoveStagePlan;
-use common_planner::plans::RenameTablePlan;
 use common_planner::plans::RevokePrivilegePlan;
 use common_planner::plans::RevokeRolePlan;
 use common_planner::plans::SetRolePlan;
 use common_planner::plans::SettingPlan;
-use common_planner::plans::ShowCreateTablePlan;
 use common_planner::plans::ShowGrantsPlan;
 use common_planner::plans::ShowRolesPlan;
-use common_planner::plans::TruncateTablePlan;
-use common_planner::plans::UndropTablePlan;
 
 use crate::optimizer::SExpr;
 use crate::plans::copy_v2::CopyPlanV2;
-use crate::plans::create_table_v2::CreateTablePlanV2;
 use crate::plans::insert::Insert;
 use crate::plans::presign::PresignPlan;
 use crate::plans::recluster_table::ReclusterTablePlan;
@@ -72,12 +61,23 @@ use crate::plans::share::RevokeShareObjectPlan;
 use crate::plans::share::ShowGrantTenantsOfSharePlan;
 use crate::plans::share::ShowObjectGrantPrivilegesPlan;
 use crate::plans::share::ShowSharesPlan;
+use crate::plans::AlterTableClusterKeyPlan;
 use crate::plans::CreateDatabasePlan;
+use crate::plans::CreateTablePlanV2;
+use crate::plans::DescribeTablePlan;
 use crate::plans::DropDatabasePlan;
+use crate::plans::DropTableClusterKeyPlan;
+use crate::plans::DropTablePlan;
+use crate::plans::ExistsTablePlan;
 use crate::plans::KillPlan;
+use crate::plans::OptimizeTablePlan;
 use crate::plans::RenameDatabasePlan;
+use crate::plans::RenameTablePlan;
 use crate::plans::ShowCreateDatabasePlan;
+use crate::plans::ShowCreateTablePlan;
+use crate::plans::TruncateTablePlan;
 use crate::plans::UndropDatabasePlan;
+use crate::plans::UndropTablePlan;
 use crate::plans::UpdatePlan;
 use crate::plans::UseDatabasePlan;
 use crate::BindContext;
